moved non-redux functions to helper methods
